roulette prizes

+++ src/main/kotlin/com/github/game/GameOfChanceResult.kt
@@ -1,5 +1,5 @@
 package com.github.game
 
 open class GameOfChanceResult {
-    var prize: Double = 0.0
+    var prize: Int = 0
 }
